Refactor: Move io utils to io module
